requirements.txt
python-dotenv~=1.2.1
fastapi~=0.128.7
pydantic~=2.12.5
openai~=2.19.0
uvicorn~=0.40.0